+++ panel-core/src/deckLayers/ArrowLayer/edge-arrow-layer.ts
@@ -4,6 +4,7 @@ import { Geometry } from 'geojson';
 import { getIconAtlasImage, iconMapping } from './arrow-atlas';
 import { toRGB4Array } from '../utils/color';
 import { ALERTING_STATES } from '../../types/defaults';
+import { getNsPrefixes } from '../../graph/utils/utils.graph';
 import { colTypes, type DeckLine, type PointFeatureProperties, type RGBAColor } from '@mapgl/panel-core/types';
 import { getEdgeArrowSize, getArrowAngle } from '@mapgl/panel-core/graph/utils';
 import { Matrix4 } from '@math.gl/core';
@@ -184,11 +185,7 @@ export const EdgeArrowLayer = (props) => {
       : {};
 
   const modelMatrix = new Matrix4();
-  const parts = srcGraphId.split('.');
-  let path = '';
-  for (let i = 0; i < parts.length; i++) {
-    path = path ? `${path}.${parts[i]}` : parts[i];
-
+  for (const path of getNsPrefixes(srcGraphId)) {
     const shift = layerShift[path];
     if (shift) {
       modelMatrix.translate([shift[0], shift[1], 0]);